Removed KafkaRecoveryConfig in bytewax.run, moved runner in its own module
